feat(forms): Support form adjusters changing fields from other sections on EntityForm (where each section is its own instance of this component).  Also add classes to animate fields appearing and disappearing.
This requires that the input formData here include all form fields on the page, not just the ones to render in this component - so to limit which ones are included here, add fieldNamesToInclude array, which, if present, controls which ones appear.

Also take setFormFields param, as a callback to modify the parent's formFields map.